Create housie_generator.rb
Each ticket shows exactly 15 numbers, and in each ticket every number from 1 to 90 is used exactly once.

A ticket is a grid of 3 rows by 9 columns. The first column contains numbers from 1 to 9; the second column numbers from 10 to 19; the third 20 to 29 and so on up until the last column, which contains numbers from 80 to 90.

Each column contains one, two or three numbers, in increasing order downwards. Each row contains exactly 5 numbers (and hence 4 blanks).